+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Protection of all component and plugin folders against direct web access
 * Update for new FOF 3 behavior regarding backend and frontend caching
+* Slightly less wonky backend downloads graph
 
 **Bug fixes**